Rework the way we add collaborators.
Rather than adding collaborators directly to the repository this app is
registered with, check whether the repository has a team called "maintainers"
and add to that instead. This makes things more manageable for the repo
maintainer, and it allows them to set the permission associated with the team
(e.g., they can use `triage`, which is so far not supported by the GitHub API).