Add "retry" command to spackbot.
Utilize the "retry" API endpoint for pipelines to re-run the jobs that
had failed during the most recent failed run of a pipeline.

It does require a new comment to listen to and a call to acquire the ID
of the most recently failed pipeline.  Use the PR ref to capture the
pipelines for that specific instance.