workers: Do not attempt to re-auth installation
According to the ghapi docs, the token should be good for one hour
from the time you request.  So instead of having the worker attempt
to re-authenticate the installation as soon as it starts (which so
far has been pretty much immediately), let's just use the token we
already put in the job metadata to communicate with the api.

If we ever see that the worker is taking more than an hour to get
kicked off or do it'w work, we can revisit this.