Block running pipeline if it would test an old commit (#70)
Block running a pipeline if it would test an old commit

If gitlab does not have the latest code from a PR on github, spackbot will no
longer allow the developer to run a pipeline.  Instead, it now responds with a
message indicating why it cannot comply and providing pointers to more
information.