Quick and dirty fixes to enable spackbot on multiple repos (#108)
* Quick fix to spackbot to enable usage on both core and packages

In lieu of a more thorough update to spackbot a quick fix to simply
enable the usage of spackbot from both of the repos maintained by the
spack organization and not an update to make spackbot configurable to
support workflows on other forks that wish to utilized the spackbot
tools. The configurations for the two repos are hard-coded, similarly to
before, for the "spack" and "spack-packages" projects in the PROJECTS
variable.

Labeling and package maintainer pinging have been dropped. These are now
handled by GitHub actions that run on pull request automatically as part
of the "triage" workflow.

ref. https://github.com/spack/spack-packages/pull/507

Style is handled by configuring the tool to run for each project. This
is hardcoded based on the project. For "spack" project the tool is the
develop version of `spack style --fix` and for the "spack-packages"
project the tool is simply `black`.

* Update to use python3 version of awscli

* Update spackbotdev redeploy script